pyproject.toml
[project]
name = "prompt-passage"
version = "0.0.2"
description = "A local proxy for LLMs, providing a unified interface for multiple models and support for identity based authentication."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity==1.23.0",
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "jq==1.10.0",
    "openai==1.79.0",
    "pydantic==2.11.4",
    "pyyaml==6.0.2",
    "uvicorn==0.34.2",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "debugpy>=1.8.14",
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "pytest==8.3.5",
    "pytest-httpx>=0.35.0",
    "ruff==0.11.10",
    "twine>=6.1.0",
    "types-pyyaml>=6.0.12.20250516",
    "types-requests>=2.32.4.20250611",
]

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.mypy]
python_version = "3.12"
files = ["core"]
strict = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = [
  "tests"
]
pythonpath = [
  "src",
  "src/prompt_passage"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
prompt-passage = "prompt_passage.cli:main"